Add DependencyManager pretty printing
This adds methods to firrtl.options.DependencyManager that enable a
user to pretty print the resulting transform order. This is extensible
via a customPrintHandling method that allows the user to define a
partial function that will be used *first* when walking the transform
order. The intended use here is to print more information or to
recurse into transforms which are not DependencyManagers.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>