Bugfix/improvement to DependencyManager scheduling
This changes the invalidation graph to explore all of the
dependencyGraph vertices (graph sum of prerequisites and dependents).
This has two effects:

  1. Fewer re-lowerings will happen as all prerequisites are
     considered when determining a schedule in a DependencyManager

  2. This enables the use of dummy/hook transforms that allow for
     conditional scheduling (more info below)

For conditional scheduling, this is the situation of correctly
scheduling an out-of-tree (custom) target (FOO) based on the
prerequisites of two other, in-tree targets (BAR, BAZ). Concretely,
given two Verilog emitters, one with minimal optimization
prerequisites (BAR) and one with full optimization
prerequisites (BAZ), the custom transform (FOO) should be scheduled
after both sets of optimizations. FOO cannot define prerequisites
since it depends on one or the other set of optimizations running.
Instead, a dummy/hook identity transform can be introduced
in-tree (QUX). All optimizations define QUX as a dependent and FOO
defines QUX as a prerequisite. Thereby, if no target defines QUX as a
prerequisite, then it is not run; if FOO is a target, then QUX will
run after either the minimal prerequisites (if BAR is a target) or
after the full prerequisites (if BAZ is a target).

This includes a test case showing this example.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>